# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Code processor",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "Login user and return sessionID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "sessionID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad json",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Register user and return usedID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "Register user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "usedID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad json",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/result/{id}": {
            "get": {
                "description": "Return result task by id",
                "summary": "Get result task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "taskID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Result",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "there's not such task",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized user",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/status/{id}": {
            "get": {
                "description": "Return status task by id",
                "summary": "Get status task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "taskID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "there's no such task",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized user",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/task": {
            "post": {
                "description": "Create task and return taskID",
                "summary": "Create task",
                "responses": {
                    "201": {
                        "description": "taskID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid json",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized user",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}